fix: 300行制限違反の最大2ファイル分割完了 - 498行・484行→300行以下
PR #537レビュー指摘の18ファイル300行超過問題に対応:

分割完了（498行→250行以下）:
- test_template_and_file_coverage.py → 2ファイル
  - test_template_system_coverage.py (テンプレート管理)
  - test_file_operations_performance_coverage.py (ファイル操作・パフォーマンス)

分割完了（484行→300行以下）:
- test_main_renderer_deep_coverage.py → 2ファイル
  - test_html_renderer_deep_coverage.py (HTMLレンダラー深度テスト)
  - test_renderer_module_integration_coverage.py (モジュール統合・エラーハンドリング)

進捗: 18ファイル→16ファイル（2ファイル削減）
残り: 16ファイルが300行超過（継続対応中）
