+++ tests/test_file_operations_performance_coverage.py
@@ -1,6 +1,6 @@
-"""Template and File Operations Coverage Tests
+"""File Operations and Performance Coverage Tests
 
-Additional Phase 2 tests for template management and file operations.
+Focus on file operations, I/O handling, and performance utilities.
 Target high-coverage modules for maximum impact.
 """
 
@@ -9,211 +9,6 @@ from pathlib import Path
 from unittest.mock import Mock, mock_open, patch
 
 import pytest
-
-
-class TestTemplateSystemCoverage:
-    """Boost template system coverage"""
-
-    def test_template_context_comprehensive(self):
-        """Test template context comprehensive functionality"""
-        try:
-            from kumihan_formatter.core.template_context import TemplateContext
-        except ImportError as e:
-            pytest.skip(f"Dependency unavailable: {type(e).__name__}: {e}")
-            return
-
-        # Test basic context creation
-        context = TemplateContext()
-        assert context is not None
-
-        # Test data operations
-        test_data = {
-            "title": "Test Document",
-            "content": "Test content",
-            "meta": {"author": "Test Author", "date": "2024"},
-        }
-
-        try:
-            context.update(test_data)
-
-            # Test data access
-            assert context.get("title") == "Test Document"
-            assert context.get("nonexistent", "default") == "default"
-
-            # Test nested access
-            if hasattr(context, "get_nested"):
-                meta_author = context.get_nested("meta.author")
-                assert meta_author == "Test Author"
-        except (AttributeError, NotImplementedError, TypeError) as e:
-            # Some methods may not be implemented
-            pytest.skip(f"Method not implemented or available: {e}")
-
-        # Test context merging
-        try:
-            additional_data = {"version": "1.0", "lang": "ja"}
-            context.merge(additional_data)
-            assert context.get("version") == "1.0"
-        except (AttributeError, NotImplementedError, TypeError, FileNotFoundError) as e:
-            pytest.skip(f"Dependency unavailable: {type(e).__name__}: {e}")
-
-        # Test context export
-        try:
-            exported = context.to_dict()
-            assert isinstance(exported, dict)
-        except (AttributeError, NotImplementedError, TypeError, FileNotFoundError) as e:
-            pytest.skip(f"Dependency unavailable: {type(e).__name__}: {e}")
-
-    def test_template_manager_comprehensive(self):
-        """Test template manager comprehensive functionality"""
-        try:
-            from kumihan_formatter.core.template_manager import TemplateManager
-        except ImportError as e:
-            pytest.skip(f"Dependency unavailable: {type(e).__name__}: {e}")
-            return
-
-        manager = TemplateManager()
-
-        # Test template loading
-        try:
-            template_names = ["default", "minimal", "detailed"]
-            for name in template_names:
-                template = manager.load_template(name)
-                assert template is not None
-        except (FileNotFoundError, AttributeError, NotImplementedError) as e:
-            # Templates may not exist
-            pytest.skip(f"Dependency unavailable: {type(e).__name__}: {e}")
-
-        # Test template rendering with context
-        try:
-            context = {"title": "Test", "content": "Content"}
-            result = manager.render("default", context)
-            assert isinstance(result, str)
-            assert len(result) > 0
-        except (AttributeError, NotImplementedError, TypeError, FileNotFoundError) as e:
-            pytest.skip(f"Dependency unavailable: {type(e).__name__}: {e}")
-
-        # Test template validation
-        try:
-            is_valid = manager.validate_template("default")
-            assert isinstance(is_valid, bool)
-        except (AttributeError, NotImplementedError, TypeError, FileNotFoundError) as e:
-            pytest.skip(f"Dependency unavailable: {type(e).__name__}: {e}")
-
-        # Test available templates
-        try:
-            available = manager.get_available_templates()
-            assert isinstance(available, (list, tuple, set))
-        except (AttributeError, NotImplementedError, TypeError, FileNotFoundError) as e:
-            pytest.skip(f"Dependency unavailable: {type(e).__name__}: {e}")
-
-    def test_template_filters_comprehensive(self):
-        """Test template filters comprehensive functionality"""
-        try:
-            from kumihan_formatter.core.template_filters import TemplateFilters
-        except ImportError as e:
-            pytest.skip(f"Dependency unavailable: {type(e).__name__}: {e}")
-            return
-
-        filters = TemplateFilters()
-
-        # Test text filters
-        test_texts = [
-            "Hello World",
-            "multiple   spaces",
-            "CamelCaseText",
-            "snake_case_text",
-            "<p>HTML content</p>",
-        ]
-
-        filter_methods = [
-            "escape_html",
-            "strip_tags",
-            "normalize_whitespace",
-            "to_uppercase",
-            "to_lowercase",
-            "to_title_case",
-        ]
-
-        for text in test_texts:
-            for filter_name in filter_methods:
-                if hasattr(filters, filter_name):
-                    try:
-                        filter_func = getattr(filters, filter_name)
-                        result = filter_func(text)
-                        assert isinstance(result, str)
-                    except (
-                        AttributeError,
-                        NotImplementedError,
-                        TypeError,
-                        ValueError,
-                    ) as e:
-                        pytest.skip(f"Dependency unavailable: {type(e).__name__}: {e}")
-
-        # Test number filters
-        test_numbers = [42, 3.14159, 1024, 0]
-        number_filters = ["format_number", "format_bytes", "format_percentage"]
-
-        for number in test_numbers:
-            for filter_name in number_filters:
-                if hasattr(filters, filter_name):
-                    try:
-                        filter_func = getattr(filters, filter_name)
-                        result = filter_func(number)
-                        assert isinstance(result, str)
-                    except (
-                        AttributeError,
-                        NotImplementedError,
-                        TypeError,
-                        ValueError,
-                    ) as e:
-                        pytest.skip(f"Dependency unavailable: {type(e).__name__}: {e}")
-
-    def test_template_selector_comprehensive(self):
-        """Test template selector comprehensive functionality"""
-        try:
-            from kumihan_formatter.core.template_selector import TemplateSelector
-        except ImportError as e:
-            pytest.skip(f"Dependency unavailable: {type(e).__name__}: {e}")
-            return
-
-        selector = TemplateSelector()
-
-        # Test template selection based on content
-        test_content_types = [
-            {"type": "document", "length": "short"},
-            {"type": "article", "length": "long"},
-            {"type": "reference", "format": "detailed"},
-            {"type": "summary", "format": "minimal"},
-        ]
-
-        for content_props in test_content_types:
-            try:
-                selected = selector.select_template(content_props)
-                assert isinstance(selected, str)
-            except (
-                AttributeError,
-                NotImplementedError,
-                TypeError,
-                ValueError,
-                FileNotFoundError,
-            ) as e:
-                pytest.skip(f"Dependency unavailable: {type(e).__name__}: {e}")
-
-        # Test auto-selection
-        sample_content = "This is a sample document with some content."
-        try:
-            auto_selected = selector.auto_select(sample_content)
-            assert isinstance(auto_selected, str)
-        except (AttributeError, NotImplementedError, TypeError, FileNotFoundError) as e:
-            pytest.skip(f"Dependency unavailable: {type(e).__name__}: {e}")
-
-        # Test template scoring
-        try:
-            candidates = ["minimal", "default", "detailed"]
-            scores = selector.score_templates(candidates, {"length": "medium"})
-            assert isinstance(scores, dict)
-        except (AttributeError, NotImplementedError, TypeError, FileNotFoundError) as e:
-            pytest.skip(f"Dependency unavailable: {type(e).__name__}: {e}")
 
 
 class TestFileOperationsCoverage:
